+++ src/components/NavItems/Assistant/Assistant.jsx
@@ -22,10 +22,10 @@ import { useTrackEvent } from "@/Hooks/useAnalytics";
 import { useSetInputFromAssistant } from "@/Hooks/useUrlOrFile";
 import AssistantIcon from "@/components/NavBar/images/navbar/assistant-icon-primary.svg";
 import AssistantCheckStatus from "@/components/NavItems/Assistant/AssistantCheckResults/AssistantCheckStatus";
+import AssistantFactCheckResult from "@/components/NavItems/Assistant/AssistantCheckResults/AssistantFactCheckResult";
 import AssistantNEResult from "@/components/NavItems/Assistant/AssistantCheckResults/AssistantNEResult";
-import AssistantSCResults from "@/components/NavItems/Assistant/AssistantCheckResults/AssistantUrlDomainAnalysisResults";
+import AssistantSCResults from "@/components/NavItems/Assistant/AssistantCheckResults/AssistantUrlDomainAnalysisResult";
 import AssistantCommentResult from "@/components/NavItems/Assistant/AssistantScrapeResults/AssistantCommentResult";
-import AssistantFactCheckResult from "@/components/NavItems/Assistant/AssistantScrapeResults/AssistantFactCheckResult";
 import AssistantLinkResult from "@/components/NavItems/Assistant/AssistantScrapeResults/AssistantLinkResult";
 import AssistantMediaResult from "@/components/NavItems/Assistant/AssistantScrapeResults/AssistantMediaResult";
 import AssistantTextResult from "@/components/NavItems/Assistant/AssistantScrapeResults/AssistantTextResult";
